+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # TBD
 
+### Features
+- The CLI now displays the list of container images currently being downloaded and validated during the Starlark
+validation step
+
 # 0.57.8
 
 # 0.57.7